requirements.txt
streamlit
tensorflow==2.11.0
keras==2.11.0
protobuf==3.20.*
numpy
Pillow
matplotlib
seaborn
scikit-learn
opencv-python
